requirements.txt
pandas
scikit-learn
joblib
sentence-transformers
transformers
torch
langchain
langchain-community
langchain-huggingface
chromadb
ollama
pymongo
python-dotenv
openai>=1.0.0

#streamlit run app.py
